pyproject.toml
[tool.poetry]
name = "jeeadv-2024"
version = "0.1.0"
description = ""
authors = ["Quantum-Codes <87054411+Quantum-Codes@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pypdf2 = "^3.0.1"
mysql-connector-python = "^9.3.0"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
